added this to the footer "social-links" width: -moz-fit-content;
and moved the image file to the same area that the "Earth.svg" is.   it never
disappeared like the others did. so it liked that path better. And I changed the name of all three images and added a 2 to the end. In case the cache was playing games with itself.  Sometimes I do that with images to correct issues.   You have to change the name of the image sometimes if they get played with too much.

 launch this and test to see if the issue still comes up.