Back to the basics
This is a fresh commit for the toc, You can have it if you want it.  It has the words Table of Contents on it as requested with no styles.   You can save it here if you want while we figure out why the renderTitle function isn't working correctly.   Maybe @Pedro can figure it out.  While i think about it .